Fix sticky_header.js
Improved sticky_header.js so that it works whenever window is resized,
even if the sticky header is already active. Did this by basing
initial_unscrolled_position on the height of the photo, rather than the
offset position of the sticky header, since the offset position
automatically becomes 0 whenever the stickiness is active and thus is
useless for comparison.