Fix weird flexbox overflow on contact.html
Restructured contact.html and removed IDs borrowed from index.html,
essentially rebuilding the contact page from scratch in order to fix the
flexbox overflow. For the emojis on the bottom, I made it so that they
appear above and below the text on portrait mode, and on either side on
landscape mode.